開始制御のマネージャーを吸収
