+++ dollsPanic/Assets/main/Script/ImageSlide.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 1c15138e2b073b949832d80ce15b9724
+timeCreated: 1528446363
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
