+++ dollsPanic/Assets/main/Script/StartProduction.prefab.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: b29e5b938c4694648b8e144880a50670
+timeCreated: 1528446366
+licenseType: Free
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
