+++ dollsPanic/Assets/script/開始演出.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: b8906ed0a8b897244aef8b6136944759
+folderAsset: yes
+timeCreated: 1528446363
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
